Alert

+++ components/FavoritesComponent.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+import { FlatList, View, Text, Alert } from 'react-native';
 
 const mapStateToProps = state => {
     return {
@@ -13,8 +13,9 @@ const mapStateToProps = state => {
         favorites: state.favorites
     };
 };
-const mapDispatchToProps = {    
-    deleteFavorite: campsiteId => (deleteFavorite(campsiteId))};
+const mapDispatchToProps = {
+    deleteFavorite: campsiteId => (deleteFavorite(campsiteId))
+};
 
 class Favorites extends Component {
 
@@ -24,12 +25,29 @@ class Favorites extends Component {
 
     render() {
         const { navigate } = this.props.navigation;
-        const renderFavoriteItem = ({item}) => {
+        const renderFavoriteItem = ({ item }) => {
             const rightButton = [
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
 
@@ -38,8 +56,8 @@ class Favorites extends Component {
                     <ListItem
                         title={item.name}
                         subtitle={item.description}
-                        leftAvatar={{source: {uri: baseUrl + item.image}}}
-                        onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
+                        leftAvatar={{ source: { uri: baseUrl + item.image } }}
+                        onPress={() => navigate('CampsiteInfo', { campsiteId: item.id })}
                     />
                 </Swipeout>
             );
